+++ src/BoardM5stackCore2.cpp
@@ -57,7 +57,7 @@ void BoardM5stackCore2::initBoard()
   Write1Byte(0x82, 0xff);
 
   M5.Axp.SetESPVoltage(3350);
-  M5.Axp.SetBusPowerMode(1); // 1 - Power from bus; 0 - Power from USB
+  M5.Axp.SetBusPowerMode((liveData->settings.boardPowerMode == 0 ? 0 : 1)); // 1 - Power from bus; 0 - Power from USB
   M5.Axp.SetLDOVoltage(2, 3300);
   M5.Axp.SetLDOVoltage(3, 2000);
   M5.Axp.SetLDOEnable(2, true);