+++ RELEASENOTES.md
@@ -3,10 +3,13 @@
 ### Next version
 
 ### v3.0.0 2024-03-16
-- no longer supported boards - TTGO-T4 and M5Stack Core1
-- new gps M5 module GNSS works with UART2 / 38400 settings
+- No longer supported boards - TTGO-T4 and M5Stack Core1
+- New gps M5 module GNSS works with UART2 / 38400 settings
 - [DEV] added support for new board M5Stack CoreS3
-- removed SIM800L module support
+- Removed SIM800L module support
+- Removed feature: Headlight reminder
+- Removed feature: Pre-drawn charging.graphs
+- Menu Board/Power mode (external/USB)
 
 ### v2.8.3 2023-12-25
 - removed debugData1/2 structures. Replaced with contribute anonymous data feature.